+++ staffdash/public/index.html
@@ -17,7 +17,8 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 		<script src="../Tools/javascript/lightClient.js"></script>
-		<script src="../Tools/javascript/lightSocket.js"></script>
+    <script src="../Tools/javascript/lightSocket.js"></script>
+    <script>src="./src/tools/app.js"</script>
     <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
@@ -42,6 +43,6 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
-    <script>src="app.js"</script>
+    <script>src="./src/tools/app.js"</script>
   </body>
 </html>
